src/vectorStoreAPI/requirements.txt
pydantic==2.6.0
fastapi==0.109.0
uvicorn==0.20.0
requests==2.27.1
python-multipart==0.0.6
python-dotenv

numpy==1.26.4
torch==2.3.1
qdrant-client==1.9.1
python-magic==0.4.27
torchvision==0.18.1
opencv-python==4.10.0.82

tqdm==4.66.4
pymongo==4.7.3
pandas==2.2.2